Create solution_dfs_recur.py
O(n) time recursive DFS solution with O(h = log(n)) space for the recursion.

Perform DFS with left and right boundary indices. If the pointers cross, then return null.

Get the value as the max value of the values in the array, and the index as the index of the max value in the array within that range passed into the function.

Then, generate a new tree node with this value, then perform DFS on the left side (not including the index) and assign to the left child of the current node, and similarly on the right side. Return the node after it is done.

Finally, return the DFS on the interval 0 to the last index of the node.