Create solution_monotonic_stack.py
O(n) iterative monotonic stack solution with O(n) potential space. This will maintain a monotonic stack.

For each node, create a node with it as the value. Then, while the stack is not empty and the top of the stack has value less than the current node, keep popping and put it as the left child of the current node. If the stack is still not empty afterwards, that means the top of the stack's right child would be the current node, since the current node is smaller than the top of the stack. Finally, append the node to the stack.

At the end, the largest value is at the bottom of the stack, with all links completed. Return this value (which should be the root of the constructed tree).