Create solution.py
O(n) time and O(1) space. 

Create a dummy node with an infinity value, and set root to the right of the dummy node. Subsequently, set a current node pointer to the dummy.

While the current's right child is not null and the current's right child value is greater than the value to be added, keep updating current to its right child. This is similar to the monotonic stack logic of traversing to the right.

Then, set a previous right value to current's right child. Meanwhile, update current's right to be a new tree node with the value to be passed in. Then, set the current right child's left child to the previous right child that was stored in the temp variable. Finally, return the dummy's right child (the root).