Create solution.py
O(n) time solution with O(n h = n log n) space for all possible paths (there are a possible n/2 paths to leaves max for a given binary tree if it is a full binary tree, and the height h is proportional to log(n), except for  cases such as those that have a linked-list like structure and is not balanced.

If the root is null, return the empty list.

Then, define a DFS function with the node, the path so far, and the remaining sum. Subtract the node value from the remaining sum, and concatenate to the temporary path list the node value. If either the left or right child exists, perform DFS with the current temporary path and the remaining sum on both. If both children don't exist and the remaining sum is 0, append this path to the result array.

Finally, perform DFS on the root with an empty list, and the target sum as the remaining sum. Return the result afterwards.