Create solution.py
O(n) time and O(n) space solution using dictionary to keep track of the sum at each root. Set the initial target sum value in the dictionary to 1.

Then, define the DFS function with a node and the root sum as arguments. Define the number of paths to be nonlocal (otherwise, it will be outside the scope). If the node is null, return.

Increment the root sum by the node value, and the number of paths by the lookup dictionary value for the root sum key. Then, increment the lookup at the root sum plus target sum by 1. Perform DFS on the left and right nodes with the updated root sum. Finally, decrement the lookup at the root sum plus target sum by 1.

Finally, perform DFS on the root passing in 0 as the root sum, and then return the updated number of paths.