Create solution.py
O(n) time complexity and O(h = log n) space from DFS recursion.

First, if one of the root'1 is null, return whether they are both null.

If the root 1 and root 2 values are different, return false.

Now, calculate recursively the non-flip trees by comparing the left subtrees of both and right subtrees of both to see if they are both equivalent. Additionally, check if the flipped versions (left and right of root 1 and root 2, respectively, as well as vice versa), are both the same.

Now, finally, return whether the flip or non-flip is true.