Create solution.py
O(mn) time solution - O(mn) space if including the 'answer' matrix that is generated, O(1) if not counting this. The extra answer matrix is created as per the problem statement. If not, there is an in-place solution as well.

Create a max per columns array that calculates the maximum for the column for each row.

Then, run through the matrix and set the values in the answer matrix to the max per column if the value is -1 otherwise just the value of the matrix at that index.

Return the answer.