Create solution.py
O(log(dividend /divisor)) solution due to dividing by divisor (height of the iteration). O(1) space. Only uses addition and subtraction (instead of multiplication).

Assign absolute value for temporary dividend and divisor variables, and set the result to 0.

While the dividend is not less than the divisor, assign the divisor to a temp variable, and multiplier to 1.

Now, while the dividend is not less than the temp variable, subtract the temp variable from the dividend, and increment the result by the multiplier. Double the multiplier by itself by adding itself to itself. Similarly, add the temp variable to itself.

If the signs of the dividend and divisor are opposite, multiply the result by -1.

Now, finally, return the minimum of the largest possible 32-bit positive integer (2^31 - 1) and the maximum of the smallest possible 32-bit negative integer (- 2^31).