Create solution.py
O(n) time and O(h = log(n)) space recursive DFS preorder traversal solution.

Create a result list. 

Define the DFS function. If the node is null, return. Otherwise, append the current node value to the result, then for each child in the node's children, perform DFS on the child.

Run DFS on the root, and return the result array.