Create solution.py
O(n) time solution where n is the length of the string input.

Define the alphabet size and the column as 0, and multiplier as 1. Start the index from the end of the string.

While the index is not out of bounds, add to the column the ord() of the character at the index, with the appropriate value calculated by differencing with the ord() of the uppercase 'A' character and adding 1 (since 'A' corresponds to 1, not 0). Then multiply this by the multiplier. Update the multiplier by multiplying by the alphabet size each iteration of the loop. Decrement the index.

Finally, return the column value.