Create solution.py
O(n) time and O(n) space solution using BFS to build up graph from the tree, then also using BFS to traverse the graph.

If k = 0, then return a list of the target's value.

Create a queue for the BFS and an adjacency list (a defaultdict) for the graph.

While the queue is not empty, pop from the queue. If the node has either a left or right child, append the pairwise relation of parent -> child and child -> parent to the adjacency list and append to the queue.

Initialize an empty result list, and created a visited set using the target as the initial value in a list. Similarly, initialize a new queue with a tuple of target and initial distance of 0.

While the queue is not empty, pop from the queue and get the node and distance. If the distance is k, append to the result its value. Otherwise, for each edge for the given node in the adjacency list, if the edge has not been visited, add this edge to the visited set, and append the edge to the queue with an incremented distance for moving 1 step.

Finally, return the result.