Create solution.py
O(n) time and O(n) space solution. n is the number of nodes (and (n-1) is the number of edges since it's a tree structure. Thus, O(V + E) where V is the number of nodes and E is the number of edges is O(V + E = n + (n-1) = 2n = n).

Now, create an adjacency matrix and populate it. Then, create a dictionary of edge counts for each node, and a leaves queue for just the leaves of the tree.

For each pair of source to neighbor in the adjacency list, append all leaves by checking the length of the neighbors equals 1. Also update the edge count for each node by the length of its neighbors.

While the leaves queue is not empty, check for the stopping condition, which is if n is 1 or 2 - these are the cases where one central node or 2 central nodes remain. Then the remaining values are the roots of the tree that minimizes the height.

Otherwise, for each of the leaves, pop from the queue and set that as the node in question, and decrement n.

For each of the neighbors of this particular node, decrement the neighbor's edge count (since the leaf node in question is being removed, which reduces the neighbors' edge counts), and if the edge count of a neighbor after this is 1, append to the leaves queue.