Create solution.py
O(n log n) time complexity solution using sorting at the end with O(n) space.

The answer list is created. The lose count dictionary is also created.

For each match, increment by 0 the lose count for the winner (this is necessary to include winners that lose 0 matches; otherwise, they won't be included in the dictionary and subsequently won't be included in the output). Similarly, increment by 1 the lose count for the loser.

Now, for each key-value pair in the dictionary, if v is 1 or 0, append to the appropriate list in the answer list. Then, for each list in the list of lists, sort them and return the answer.