Create solution.py
O(1) time complexity for lock() and unlock() functions, and O(n) time for constructor and the upgrade() function. O(n) space for the parent array, children dictionary, and the locked array.

Initialize for each parent its children in the constructor once the parent array is passed in.

For lock() and unlock(), check the condition and see if possible. Then, set None -> user or user -> None, vice versa, depending on the operation.

For the upgrade() function, set the current index i to the num argument. While this index is not equal to -1 (meaning the root), if the i is not locked (the node itself or its parent), keep updating the index i to its parent. If not, return false.

Afterwards, keep track of the locked count, and initialize a queue with the num (node to start upgrade from). This is a BFS algorithm to go through the children. Then, while the queue is not empty, pop from the queue, and if it is locked, unlock it. Increment the locked count. Now, extend the queue by the children of the node popped from the queue.

Now, after the operation through the queue is done, if the lock count is still greater than 0, lock the given node. Finally, return whether there were any locked descendants.