Create solution_sliding_window.py
O(n) time and O(1) space sliding window solution.

Initialize Alice and Bob's scores to 0, and the left and right pointers to 0.

While r does not cross the threshold of the colors array, if the left and right pointer values are not the same, set the left pointer to the right pointer.

Calculate the extra characters as the number of indices between the left and right pointers (r - l + 1) and subtract 2 (since the 2 endpoints do not count towards the score).

If the extra characters are greater than 0, if the rightmost is the character A, then increment Alice's score, and if it is a B, then increment Bob's score.

Return whether Alice's score is greater than Bob's score.  If they are equal, Bob wins, since with no moves left, Alice (who goes first) loses.