Create solution.py
Gradient descent solution with O(2n^2 = n^2) time complexity, since 2 is the number of coordinates, which is x and y. Space complexity is O(1) for storing the current x and y values and distance.

If n is 1, return 0.

Define the distance function, which calculates the sum of all positions from the input x and y coordinates.

Next, set initial x and y coordinates to (0, 0), and calculate the current distance based on this initial point. Set the step to 1.

While the step is greater than the threshold of 10^(-6), set f to false. For each delta x and y directions (+/-1 for both x and y directions), calculate the steps x and y and the distance. If the distance is less, update the distance and current x and y. Set f to true.

If f is not true, then divided the step by 10 (i.e., step-size is decreased).

Finally, return the current distance.