Create solution.py
O(m + n) time solution, where m and n are the lengths of lists list1 and list2. O(1) space solution.

Set the current pointer to the head of list1 and index i to 0.

While the index is less than a (right before the a-th index node), keep advancing the current pointer and incrementing the index.

Set this curr node to head.

Next, while the index reaches b (the index that comes next to the b-th node), advance the current pointer and increment the index.

Set the next of head to list2.

While list2 has a next value, keep advancing the list2 pointer. This stops at the last node of list2 (before the null pointer). Set the next pointer of the last element of list2 to the current node, then return list1.