Create solution.py
O(n) time and O(1) space solution where n is the length of students and sandwiches arrays (they are the same length). Only an array of size 2 is needed for space.

First, initialize the result to the length of the students array.

Next, calculate the count of each type corresponding to the indices 0 and 1. The sum of students array represents the number of students who want type 1 sandwiches, whereas the length of the array - this sum is the number of students who want type 0.

Now, for each of the sandwiches, if the count of that type of sandwich index in the array is greater than 0, decrement the result and the count array at that index. Otherwise, immediately return the result.

Finally, return the result after running through the array.