Create solution.py
O(2^n) time solution with O(2^n) space complexity due to exponential structure of enumerating all trees.

First, notice if n is even, any tree has no possible full binary trees.

Now, populate a memoization DP table with the base case of n = 0 being an empty list, and n = 1 with a single tree node.

Define the backtracking function and check if the value is in DP, then return the value.

Initialize a result array, and iterate through a left index for the left subtree through n. Calculate the corresponding r value . If either r or l is not odd, go to the next iteration.

Then, calculate the backtrack of the left and right subtrees, and for each of the combinations of left and right subtrees, append to the result array a new treenode with the left and right subtrees. Add this value to the DP array. and return the value.

Finally, run backtrack on n and return.