Create solution_sorted.py
O(n log n) time and O(n) space solution from sorted arrays.

Initialize the best value and last x values, as well as the total. Simiarly, initialize the delta values for the next and previous scores.

Go through the zipped nums and cost values and get the total costs and deltas by looking at the costs.

Then, go through the sorted zipped lists with calculating the delta of the x value, and decrement the next delta and increment the next delta for the total, and similarly for the deltas themselves. Afterwards, calculate the minimum best value and update the last x.

Return the best value.