Create solution_gd.py
O(n log n) gradient-descent like solution by calculating left and right side and appropriately adjusting the values. O(1) space.

Define the calculate cost function by calculating the absolute difference times cost.

The result needs to be one of the values in the array. Thus, set the left bound to the min value of the array and right bound as the max value in the array.

While the bound pointers don't cross, update the target.

Calculate the current cost (using current target), left cost (by subtracting 1 from the current target), and right cost (by adding 1 to the current target).

If both the left and right costs are greater than or equal to the current cost, the current cost is the minimum, so return. Otherwise, if the left cost is less, move the right bound to the target - 1 value (since the left direction has a lower value). Otherwise, it is the case where the right side has a lwoer value, so move the left bound to the target + 1 value (the right side has a lower value).

Finally, return the lower bound.