Create solution.py
O(log n) time solution.

Set the result to 0 initially.

Now, for each of the 32-bits in the integer, get the i-th bit from the right by shifting the left value by i. If the bit is 0, go to the next iteration.

Calculate the remainder by modding the left value by the binary value that has 1 shifted (i+1) times, which is the next power of 2 above the current value. Then, difference this value against the remainder value, and if the  right - left is less than this difference, perform an or operation with 1 shifted i times, which will give whether this value is 0 or not.

Return the result.