Create solution.py
O(n) time and space solution using a 3-pointer sliding window and hashmap.

Create a hashmap of counts of each values encountered, along with initializing the result to 0, and the near and far left pointers to index 0.

For right pointer until the end of the array, increment the count in the dictionary by 1. If the length of the count dictionary becomes greater than k, decrement the value at the near left pointer, and if this becomes 0, pop that key. Increment the left pointer subsequently, and set the far left pointer to the near left pointer.

Similarly, if the count of the near pointer element in the dictionary is greater than 1, that means the left near pointer can be shifted, so decrement this count from the dictionary, then increment the near left pointer.

If the length of the count dictionary becomes k, add the length of the left pointer range (near to far).

Finally, return the result.