Create solution.py
O(n) time and O(1) space solution by modifying the nums array in place.

First, get the initial length of the nums array. 

Then, while n is greater than 1, go through the array from the beginning to n in increments of 2. Get the 2 contiguous pair of numbers and the divisor, which is calculated to see if it is odd or even. Then the minimum is set for odd, max for even at that index.

The n is divided by 2 each iteration.

Eventually, the first element of nums will have the value in question.