Create solution.py
O(n) time and O(1) space solution.

Create a set of sums. If the length of the array is 2, return false.

Calculate the current sum using the first 2 elements of the array, and add this to the set of sums.

For the range 1 up to right before the last element of the array, subtract the previous index element and add the next index element from the current sum, and if this exists in the set of sums, return true. Add the current sum to the set of sums.

If the loop exits, return false.