Create solution.py
O(n) time and space solution using a monotonic stack.

First, initialize the modulus value, result array, and the stack.

For the index and number pair in the array, while the stack is not empty and the current number is less than the top value of the stack, pop the index/number pair. Set the left value which includes all subarrays with minimum value at index j as the index difference between current index j and the top of the stack's value at the current location if the stack is not empty. Otherwise, it will just be the current index j + 1, since it would look at the previous value. The right side of j is i - j, since the index i and index j difference is the number of values which have the value at j as the minimum.

Update the result with the possible combinations (m * left * right) modded with the result.

Append the current i and n to the stack.

Now, again go through the stack if it is not empty, and get the index/number pair from the stack from bottom-to-top.

Calculate the left side by getting index j and finding the difference with the index just left of i if i is greater than 0, otherwise it is j + 1 as before. The right side is the end of the array, len(arr) minus j, which is the number of elements to the left and right of the current j. Update the result similarly by multiplying the number of combinations n * left * right and adding to the current result and modding.

Return the result.