Create solution_opt.py
O(n) time and space solution using a monotonic stack, but with smaller constant (no need to run through stack again).

Use a sandwiched array by book-ending -infinity values to the array. This makes it so that the stack gets emptied and there is no need to run through the stack for any remaining values again.