Create solution.py
Use a monotonic stack and prefix sum of prefix sum array. O(n) time and O(n) space solution.

Initialize n as the length of the strength array, the mod number, and the left and right index arrays initialized with -1 and n both n times, resepctively. Also, put a 0 in front of the prefix array, then another 0 in front of the prefix of prefix array using the above array. Also, initialize a stack and the result to 0.

Go through the strength array and while the stack is not empty and has value greater than or equal to the value, pop from the array. If the stack is not empty, set -1 at the left array index i. Otherwise, set its value to the index at that stack. Append the value, index pair to the stack.

Similarly, re-initialize an empty stack and go through from the right-end of the strength array and perform the same operation for the right index array.

Next, go through the length of the strength array, and get the left and right lengths for the minimum values. Now, initialize the subtotal to 0, and add the multiplication of the left length and difference in prefix of prefix arrays and subtract the right length time the prefix of prefix arrays. Afterwards, add the subtotal times the strength at that index and the mod it.

Finally, return the mod of the result.