Create solution.py
O(nm) solution where n is n and m is the possible number of characters in the length of the string.

First, decrement n by 1 (first step) since x will be assigned to the default value of '1'.

Then, define the run-length encoding (RLE) recurse function. Set the string array to empty, and group the string by values and counts. Append to the string array the string of the length of the list of the string itself and the value (this will groupby each character the number of times it occurs and the character appended for each pair. Finally, return the joined string.

Now, iterate through the range n and perform this n-1 (since n was adjusted above) times.

Return this value.