Create solution.py
O(n) time solution and O(1) space. The trick is to just assume local inversions, and if the difference in local inversions is larger than 1, then the number of global and local inversions will not be equal.

Go through the array, and check if the value of the nums array at that index is not the index (since the values range from 0 to n-1 for an n-length array). Then, check if the value at the i+1 index and i-th index are swapped. If so, then swap. Otherwise, this means the local and global inversions are not the same, and return false.

Finally, outside the loop, return true since the global = local for the inversions.