Create solution.py
O(n) time and O(n) space solution using a stack.

Initialize a stack and the index to 0.

While the index is less than the length of the string, if the stack is not empty and the top of the stack is not equal to the current character in the string but the lower case of the value at the top of the stack equals the lower case of the current character in the string (which means one value is the uppercase, the other the lowercase), pop the value from the stack. Otherwise, append the value to the stack. Finally, increment the index.

Return the joined string of the characters in the stack. In a real stack, it would need to be iterated from the top down, and reversed to yield the correct string, but this method is done single-handedly for convenience.