Create solution.py
O(n) time complexity using variation on bucket sort - no actual sorting involved, just bucketization. O(1440) = O(1) space using 60 * 24 = 1440 minute time buckets, and another array to store all the minutes available.

First, calculate the total minutes in a day (1440). Do a check to see if the number of elements in the timePoints array is greater than 1440, then return 0 (at least one time is a duplicate).

Define the minutify function, which converts the string into minutes up to 1440 as the maximum.

Create a time bucket array with 1440 buckets with 0 as the initial values.

For each time in timePoints, minutify it, then if the value in that time bucket is not 0, return 0 (since there is a duplicate time point). Otherwise, set the value at that minute to 1.

Now, create a minutes array that goes through the range of 1440 minutes, and if the value in the time buckets array for that minute is 1, add that value to the minutes array.

Finally, return the minimum of the pairwise difference between the i-th and (i-1)-th index for the values in the minutes array calculated right above.