Create solution.py
O(n) time and space solution using two monotonic stacks.

Similar to the  sum of subarray minimums, but use two passes for the monotonic increasing and decreasing stacks.

Initialize the nums array by augmenting it with -infinity at the bookends. Then, initialize an empty stack.

For the enumerated pair of index and number, while the stack is not empty and the stack's value is greater than n, pop from the stack. Calculate the left and right sides of the current index, and multiply these by the value popped and subtract from the result. Append the tuple of index/number pair to the stack.

Now, re-assign the bookend-augmented nums array with infinity on both ends this time, and re-initialize the stack to empty.

Now, analogous to the above, but now checking if the value at the top of the stack is less than the number, pop and calculate the left and right sides. Multiply these with the value, and add to the result. Finally, append the index/number pair to the stack.

Finally, return the result.