Create solution.py
O(n) time and space solution using monotonic stack.

Set the max height to 0 (since this is under the minimum value of height input possible). Initialize a stack.

Going through the array in reverse order, if the current height is greater than the max height, append to the stack (keeps the monotonicity condition). Then, update the max height.

Finally, return the reversed stack.