Create solution_stack.py
O(n) time and O(n) space solution using monotonic stack.

Initialize the result to 0 and the stack.

Go through each of the price/index pair. Set the counter to 0 and the index to 1.

If the stack is not empty and the difference between the top of the stack and the price is not 1, keep popping from the stack and add the index to the counter, and increment the index.

Finally, add the updated counter to the result.

If the stack is empty or the stack exists with the difference between the top of the stack and the price equal to 1, then append the index/price pair to the stack and skip to the next iteration.

Finally, set the counter to 0 again and the index to 1.

Run through the stack and keep popping, adding the index to the counter and incrementing the index.

Add the counter to the result and return the result.