Create solution.py
O(n^2) time and O(n^2) space solution from creating the result array.

First, initialize the length of the array, the modulus value, and the result array.

For each i in the length of the array, set a running current sum to 0. Then, for each of the j from the range of i to the end, add to the running sum and append to the result array.

Sort the result array (O(n log n)).

Finally, return the sum from the appropriate indices left to right and mod by the modulus.