Create solution.py
O(n) time methods for both encode() and decode() methods, where n is the string length.

First, create a class variable delim to be the hash character '#' (arbitrary choice) as the delimiter to be used.

Then, in encode(), encode the length of each string in the list in string form, append the delimiter and then the string, and return this at the end.

In decode(), initialize a result array and the index pointer 0. While the index is less than the length of the input string, set it equal to j. While the character is not the delimiter character, keep incrementing j.

Set the string length to the number that comes before the delimiter, which is from i to j-1. Then, append the result as the string whose character starts from the j + 1 index until the length just calculated above. Finally, set i to the j + 1 + the length of the string above to start the next iteration.

Finally, return the resulting list of strings.