Create solution.py
O(n) time and O(1) space solution.

The condition to check is pairwise and see for odd indices if the previous value is less than or equal, and for even indices if the previous value is greater than or equal. Then, there is no need to swap.

For each value starting from index 1 to the end of the array, if the index is odd but its previous value is greater, or if the index is even and the previous value is less, then the number at the current index is swapped with the number at the previous index. Then, continue the next iteration.

Finally, return the in-place wiggle-sorted array.