Create solution.py
O(n) time and O(n) space solution using hashmap.

First, initialize the difference variable that counts the number of ones - number of zeroes and the result. Also, initialize the hashmap of differences of ones over zeroes and the index at which the first such difference in counts is encountered.

For each index/number pair, if n is 0 decrement the number of ones over zeroes, and if n is 1 increment it.

If this difference in counts is not already in the dictionary, set the difference and the first index at which this occurs and add it into the dictionary.

If the number of ones over zeroes (the difference in counts of ones and zeroes) is 0, then update the result to the index + 1 (since the count of elements at this point is index + 1). Otherwise, find the index from the dictionary, and update the result to the maximum of the current result and i minus the index (this is the length of the ones and zeroes that match).

Return the result.