Create solution_dict.py
O(n) time and O(1) space solution using hashmap (there are only 26 lowercase alphabet characters for the keys of the dictionary, so constant space).

Create a counter dictionary for the original string s.

Next, iterate through each character in string t. If the character does not exist in the original characters dictionary, return this character. Otherwise, decrement the value from the dictionary. After decrementing, if the value for the character is less than 0, this means this is an additional occurrence of that particular character in string t, so return that character. This loop will always return a value within the iterations, so there is no need to return a value outside the loop.