Create solution_ascii.py
O(n) time solution and O(1) space solution. Create a character sum variable for the ASCII number values and initialize this to 0.

For each character in string t, sum the ord() of that character. For each character in string s, subtract the ord() of that character from the same sum.

Finally, return the chr() of that number (which will convert the corresponding ASCII number to the character itself).