Create solution_xor.py
O(n) time and O(1) space solution using bit manipulation by XOR.

Initialize a result value to 0.

For each character in s, convert the character to ASCII integer equivalent and XOR to the result (which is initially 0). This needs to be done since Python cannot XOR characters (only integers). Then, for each character in t, convert the character to the ASCII integer equivalent and XOR to the result.

Finally, return the character of the resulting XOR-ed value (this will contain the character whose difference is the missing character).