Create solution.py
O(n) time and O(n) space solution from recursion using postorder DFS.

First, create two boolean variables to check if p exists and q exists. 

Then, define a DFS function in postorder. If the node is null return null. Check the left and right side using DFS. If either the node is p or q, appropriately set the boolean variables, and return the node. If left and right are not null, return the node. Otherwise, return left or right.

Finally, run DFS on the root node, and return the result if both p and q exists, otherwise return null.