Create solution.py
O(n) time and O(1) space solution using two iterations using copy traversal.

Set copies of the initial p and q nodes. Then, while the two copy nodes are not the same, if the nodes is not null, set to its parent. Otherwise, switch to either p (for q copy) or q (for p copy). This is essentially traversing p and q's paths to their LCA, so the total traversal length will be len(p -> LCA) + len(q -> LCA).

Finally, return either p copy or q copy (this will be the same node when the loop terminates).