Create solution.py
O(n) time and O(n) space using recursive DFS and a set to transform the input list of nodes.

First, convert the list of nodes to a set of nodes to have constant lookup.

Then, define the DFS function as in LCA of Binary Tree I question. 

If the node is null, return null. If the node is in the nodes set, return the node (no need to look for other nodes in the set, just encounter the first one at the top).

Get the left and right subtree nodes. If the left and right are not null, return the node. Otherwise, return either the left or right node.

Finally, run DFS on the root and return this.