Create solution.py
O(n) time and O(n) space solution using recursive DFS.

First, if the root is null, return null.

Initialize the result to null and the max depth to -1.

Then, define the DFS solution with parameters node and depth.

If the node is null, return -1. If the left and right children doesn't exist, this is a leaf node. Thus, check if the current depth is greater than max depth. Then, update the result with the current node and the max depth with current depth, and return the depth. Otherwise, run DFS on the left and right children with an incremented depth. If the left, right, and max depth are all the same, then set the result to the current node. Finally, return the max of the left and right depths.

Perform DFS on the root node with depth 0, and return the result.