Create solution_bfs.py
O(n) time and O(n) space BFS solution using a queue.

First, if the depth is 1, return a node with the new value and its left child as the current root. 

Initialize the queue to contain the root and the current depth to 1.

While the queue is not empty, if the current depth is equal to depth - 1, break out of the loop. 

Otherwise, get the length of the queue and iterate through it. Pop the node and append its left and right children if they exist. Increment the current depth afterwards.

For the nodes remaining in the queue after reaching the desired depth (right above the depth to insert at), for each node remaining in the queue, get the node's current left and right subtrees. Then, set the node's left and right children to the value to be added, with its child as the previous left subtree and right subtree).

Finally, return the root.