Create solution.py
O(n) time solution using string functions.

First, make the banned list a set. Initialize the count dictionary of each word count. Also, initialize the max count and the word with the max count to 0 and an empty string, respectively.

Next, filter the word by using the regular expression split on all spaces and delimiter characters. Then, for each word, convert them to lower case.

For each word in the list, if the word is not in the banned words set, increment the count in the count dictionary. If the count becomes larger than the max count, set the max count and the word with max count appropriately.

Finally, return the word with the max count.