Create solution_graph.py
O(n) time and space solution using a graph formulation. There is O(n) space and time used for building the graph, and O(n) time and space for the BFS when looking for the distance between p and q.

First, if p and q are equal, return 0.

Initialize the dictionary for the adjacency matrix. Then, initialize the queue that will append the nodes to traverse in the graph with the root node.

While the queue is not empty, pop from the queue. If the left child of the node exists, append the edge by adding the node from parent to child and child to parent, and append to the queue. Similarly, if the right child exists, then add the edge from parent to the right child, the right child to the parent, and append the right child to the queue.

Now, initialize the queue to the tuple of p and steps taken of 0. Also create a visited set of visited nodes with p initially.

While the queue is not empty, get the node and number of steps taken pair by popping from the queue. If the node equals q, return the number of steps. otherwise, for each of the neighbors for this node, if the neighbor has not been visited, add the neighbor to the visited set, and append the neighbor to the queue with number of steps taken incremented by 1. This will prevent cycles by having a visited set to keep track of which nodes have been visited during BFS. This loop will always terminate, since the node q will always exist.