Create solution_hm.py
O(nm) time solution in the length of words dictionary n and avg length of words m.

Create an array of size 26 for the 26 lowercase letters, and offset as the ord() of the first letter. Also, initialize the result array.

For the index in the length of words, get a counter dictionary of the words. Then, for each index in the array of size 26, if the character is not in the word dictionary, set that point in the index to 0. Otherwise, set it to the counter dictionary's value if it is the first word being analyzed, otherwise the minimum of the current value in the array and the count in the dictionary.

Now, for each value in the character boolean, if the value is not null, for the value in that index, append the character that number of times.

Finally, return the result.