Create solution_bitmin
O(n) time and O(1) space solution using bitwise XOR for initial value and bitwise-min value  for the other values. Get the bitwise or to persist the values into the char count array, and find the bitwise minimum value otherwise.

Then, append to the result the character number for each value.