Create solution.py
O(n) time solution for each binary digit. 

Initialize the result and index to 0.

While the number is not 0, if the number is not 0, append to the result the i index bitshifted by 1 to the left. Increment the index and shift the number to the right by 1 in order to get the next value.

Return the result.